+++ python/spider-py/src/spider_py/client/__init__.py
@@ -1,7 +1,7 @@
 """Spider python client."""
 
 from .task import TaskContext
-from .taskgraph import chain, group, TaskGraph
+from .task_graph import chain, group, TaskGraph
 
 __all__ = [
     "TaskContext",
